version2/requirements.txt
requests
python-telegram-bot
moodleapi

